libastfri-output warnings cleaning

+++ libastfri-output/visitors/PseudocodeVisitor.cpp
@@ -1,3 +1,4 @@
+#include <cstddef>
 #include <libastfri/structures/Declaration.hpp>
 #include <libastfri/structures/Expression.hpp>
 #include <libastfri/structures/Statement.hpp>
@@ -164,7 +165,7 @@ void PseudocodeVisitor::Visit(lsfs::FunctionDefinition const& functionDef)
     this->getFormatter().print(" ");
     this->getFormatter().printIdentifier(functionDef.name);
     this->getFormatter().print("(");
-    for (int i = 0; i < functionDef.parameters.size(); i++)
+    for (unsigned long i = 0; i < functionDef.parameters.size(); i++)
     {
         functionDef.parameters[i]->accept(*this);
         if (i < functionDef.parameters.size() - 1)
@@ -240,7 +241,7 @@ void PseudocodeVisitor::Visit(lsfs::FunctionCallExpression const& expr)
     // this->getFormatter().print(expr.functionName + "(");
     this->getFormatter().printIdentifier(expr.functionName);
     this->getFormatter().print("(");
-    for (int i = 0; i < expr.arguments.size(); i++)
+    for (unsigned long i = 0; i < expr.arguments.size(); i++)
     {
         expr.arguments[i]->accept(*this);
         if (i < expr.arguments.size() - 1)
